+++ src/tables/cmap/format10.rs
@@ -15,3 +15,18 @@ pub fn parse(data: &[u8], code_point: u32) -> Option<u16> {
     let idx = code_point.checked_sub(first_code_point)?;
     glyphs.get(idx)
 }
+
+pub fn codepoints(data: &[u8], mut f: impl FnMut(u32)) {
+    let mut s = Stream::new(data);
+    s.skip::<u16>(); // format
+    s.skip::<u16>(); // reserved
+    s.skip::<u32>(); // length
+    s.skip::<u32>(); // language
+    let first_code_point: u32 = try_opt_or!(s.read(), ());
+    let count: u32 = try_opt_or!(s.read(), ());
+
+    for i in 0..count {
+        let code_point = try_opt_or!(first_code_point.checked_add(i), ());
+        f(code_point);
+    }
+}